refactor(ReadyToUse): Move to Home directory
